qna/requirement.txt
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.10.2
alto-xml==0.0.5
anyio==3.6.2
async-timeout==4.0.3
attrs==22.2.0
Authlib==1.2.0
azure-ai-formrecognizer==3.2.1
azure-common==1.1.28
azure-core==1.26.3
backoff==2.2.1
beautifulsoup4==4.12.2
bs4==0.0.1
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
cmake==3.26.0
contourpy==1.0.7
cryptography==39.0.2
cycler==0.11.0
databricks-cli==0.17.5
datasets==2.14.4
decorator==5.1.1
dill==0.3.7
dnspython==2.3.0
docker==6.0.1
docopt==0.6.2
elasticsearch==7.17.9
entrypoints==0.4
evaluate==0.4.0
faiss-cpu==1.7.2
faiss-gpu==1.7.2
farm-haystack==1.14.0
fastapi==0.78.0
ffmpeg==1.4
filelock==3.10.0
Flask==2.2.3
fonttools==4.39.2
frontmatter==3.0.7
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
grpcio==1.47.5
grpcio-tools==1.47.5
gunicorn==20.1.0
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
inflect==6.0.2
isodate==0.6.1
itsdangerous==2.1.2
jarowinkler==1.2.3
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
langdetect==1.0.9
lit==16.0.0
llvmlite==0.39.1
loguru==0.6.0
lxml==4.9.2
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.1
memory-profiler==0.61.0
mlflow==2.2.2
mmh3==3.0.0
monotonic==1.6
more-itertools==9.1.0
mpmath==1.3.0
msrest==0.7.1
multidict==6.0.4
multiprocess==0.70.15
networkx==3.0
nltk==3.8.1
num2words==0.5.12
numba==0.56.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opencv-python-headless==4.8.0.76
opensearch-py==2.2.0
packaging==23.0
pandas==1.5.3
pdf2image==1.16.3
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
posthog==2.4.0
protobuf==3.20.2
psutil==5.9.7
psycopg2-binary==2.9.5
pyarrow==11.0.0
pycparser==2.21
pydantic==1.10.6
PyJWT==2.6.0
pymilvus==2.2.3
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.7.1
PyYAML==5.1
quantulum3==0.8.1
querystring-parser==1.2.4
rank-bm25==0.2.2
rapidfuzz==2.7.0
rdflib==6.3.1
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
responses==0.18.0
rouge==1.0.1
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.97
seqeval==1.2.2
shap==0.41.0
six==1.16.0
slicer==0.0.7
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.5
SPARQLWrapper==2.0.0
SQLAlchemy==1.4.47
SQLAlchemy-Utils==0.40.0
sqlparse==0.4.3
starlette==0.19.1
svgwrite==1.4.3
sympy==1.11.1
tabulate==0.9.0
tenacity==8.2.2
tesseract-ocr-utils==0.0.6
threadpoolctl==3.1.0
tika==2.6.0
tiktoken==0.3.2
tokenizers==0.13.2
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
transformers==4.25.1
triton==2.0.0
typing_extensions==4.5.0
ujson==5.4.0
urllib3==1.26.15
validators==0.19.0
weaviate-client==3.10.0
websocket-client==1.5.1
Werkzeug==2.2.3
xxhash==3.3.0
yarl==1.9.2
zipp==3.15.0
